Fetch All Versions from go.dev (#6)
It's possible to get the versions of Go from go.dev. I figured it out from reading the website src, but since it isn't well documented, I assume that the API is subject to change without notice. It's best not to make a hard dependency on it. I'll keep a backup way to resolve the download URLs.

* get the shasums and versions from go.dev
* fix new discovery test
* handle installs as root